♻️ Table data moved to data folder
