+++ libraries/core-react/stories/components/Tooltip.stories.tsx
@@ -1,6 +1,8 @@
 import React, { useState, useRef } from 'react'
 import styled from 'styled-components'
 import { Tooltip, TooltipProps, Typography, Button, Table } from '@components'
+import { data, columns } from './helpers/data'
+import { toCellValues } from './hooks/toCellValues'
 import { Story, Meta } from '@storybook/react'
 
 const Body = styled.div`
@@ -134,104 +136,6 @@ export const WithDisabledElements: Story<TooltipProps> = () => {
     </Body>
   )
 }
-
-type Data = {
-  number: string
-  description: string
-  origin: string
-  price: number
-}
-
-const data: Data[] = [
-  {
-    number: '123-456',
-    description: 'Pears',
-    origin: 'Europe',
-    price: 1.5,
-  },
-  {
-    number: '234-567',
-    description: 'Apples',
-    origin: 'Africa',
-    price: 1.2,
-  },
-  {
-    number: '45-6789',
-    description: 'Oranges',
-    origin: 'South America',
-    price: 1.8,
-  },
-  {
-    number: '67-890',
-    description: 'Kiwi',
-    origin: 'Australia',
-    price: 2.1,
-  },
-  {
-    number: '89-012',
-    description: 'Mango',
-    origin: 'South Africa',
-    price: 2.5,
-  },
-  {
-    number: '89-012',
-    description: 'Pineapple',
-    origin: 'Paraguay',
-    price: 1.9,
-  },
-  {
-    number: '89-012',
-    description: 'Pomegranate',
-    origin: 'Persia',
-    price: 4.5,
-  },
-]
-
-type Column = {
-  name: string | React.ReactNode
-  accessor: string
-  sortDirection?: 'ascending' | 'descending' | 'none'
-  isSorted?: boolean
-}
-
-const columns: Column[] = [
-  {
-    name: 'Item nr',
-    accessor: 'number',
-    sortDirection: 'none',
-  },
-  {
-    name: 'Description',
-    accessor: 'description',
-    sortDirection: 'none',
-  },
-  {
-    name: 'Origin',
-    accessor: 'origin',
-    sortDirection: 'none',
-  },
-  {
-    name: (
-      <>
-        Price &nbsp;
-        <Typography group="input" variant="label" color="currentColor">
-          ($)
-        </Typography>
-      </>
-    ),
-    accessor: 'price',
-    sortDirection: 'none',
-  },
-]
-
-const toCellValues = (data: Data[], columns: Column[]) =>
-  data.map((item) =>
-    columns.map((column) =>
-      typeof item[column.accessor] !== 'undefined'
-        ? (item[column.accessor] as string)
-        : '',
-    ),
-  )
 
 export const TableCellsWithTooltip: Story<TooltipProps> = () => {
   const cellValues = toCellValues(data, columns)
