--- libraries/core-react/stories/components/helpers/data.ts
@@ -1,13 +1,0 @@
-export const items = [
-  'Oslo',
-  'Rogaland',
-  'Møre og Romsdal',
-  'Nordland',
-  'Viken',
-  'Innlandet',
-  'Vestfold og Telemark',
-  'Agder',
-  'Vestland',
-  'Trøndelag',
-  'Troms og Finnmark',
-]
