+++ libraries/core-react/stories/components/helpers/data.tsx
@@ -1,0 +1,106 @@
+import React from 'react'
+import { Typography } from '@components'
+
+export const items = [
+  'Oslo',
+  'Rogaland',
+  'Møre og Romsdal',
+  'Nordland',
+  'Viken',
+  'Innlandet',
+  'Vestfold og Telemark',
+  'Agder',
+  'Vestland',
+  'Trøndelag',
+  'Troms og Finnmark',
+]
+
+export type Data = {
+  number: string
+  description: string
+  origin: string
+  price: number
+}
+
+export const data: Data[] = [
+  {
+    number: '123-456',
+    description: 'Pears',
+    origin: 'Europe',
+    price: 1.5,
+  },
+  {
+    number: '234-567',
+    description: 'Apples',
+    origin: 'Africa',
+    price: 1.2,
+  },
+  {
+    number: '45-6789',
+    description: 'Oranges',
+    origin: 'South America',
+    price: 1.8,
+  },
+  {
+    number: '67-890',
+    description: 'Kiwi',
+    origin: 'Australia',
+    price: 2.1,
+  },
+  {
+    number: '89-012',
+    description: 'Mango',
+    origin: 'South Africa',
+    price: 2.5,
+  },
+  {
+    number: '89-012',
+    description: 'Pineapple',
+    origin: 'Paraguay',
+    price: 1.9,
+  },
+  {
+    number: '89-012',
+    description: 'Pomegranate',
+    origin: 'Persia',
+    price: 4.5,
+  },
+]
+
+export type SortDirection = 'ascending' | 'descending' | 'none'
+export type Column = {
+  name: string | React.ReactNode
+  accessor: string
+  sortDirection?: SortDirection
+  isSorted?: boolean
+}
+
+export const columns: Column[] = [
+  {
+    name: 'Item nr',
+    accessor: 'number',
+    sortDirection: 'none',
+  },
+  {
+    name: 'Description',
+    accessor: 'description',
+    sortDirection: 'none',
+  },
+  {
+    name: 'Origin',
+    accessor: 'origin',
+    sortDirection: 'none',
+  },
+  {
+    name: (
+      <>
+        Price &nbsp;
+        <Typography group="input" variant="label" color="currentColor">
+          ($)
+        </Typography>
+      </>
+    ),
+    accessor: 'price',
+    sortDirection: 'none',
+  },
+]
